Add a EECategory file
