+++ ElevateEngine/src/ElevateEngine/Core/EEObject.cpp
@@ -1,0 +1,13 @@
+#include "EEObject.h"
+
+namespace Elevate
+{
+	ByteBuffer Elevate::EEObject::Serialize() const
+	{
+		return ByteBuffer();
+	}
+
+	void EEObject::Deserialize(const ByteBuffer& data)
+	{
+	}
+}
